Plot Service (Python)/PythonApp/requirements.txt
APScheduler==3.10.3
asgiref==3.7.2
azure-core==1.28.0
azure-identity==1.13.0
azure-storage-blob==12.17.0
blinker==1.6.2
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
click==8.1.4
colorama==0.4.6
contourpy==1.1.0
cryptography==41.0.2
cycler==0.11.0
Django==4.2.3
django-background-tasks==1.2.5
django-compat==1.0.15
django-crontab==0.7.1
django-scheduler==0.10.1
djangorestframework==3.14.0
Flask==2.3.2
fonttools==4.41.1
greenlet==2.0.2
icalendar==5.0.7
idna==3.4
isodate==0.6.1
itsdangerous==2.1.2
Jinja2==3.1.2
kiwisolver==1.4.4
MarkupSafe==2.1.3
matplotlib==3.7.2
msal==1.23.0
msal-extensions==1.0.0
nltk==3.2.4
numpy==1.25.1
packaging==23.1
pandas==2.0.3
Pillow==10.0.0
portalocker==2.7.0
preprocessing==0.1.13
psycopg2==2.9.6
pyarrow==12.0.1
pycparser==2.21
PyJWT==2.8.0
pymsteams==0.2.2
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2023.3
pywin32==306
requests==2.31.0
six==1.16.0
sphinx-rtd-theme==0.2.4
SQLAlchemy==2.0.18
sqlparse==0.4.4
timestamp==0.0.1
typing_extensions==4.7.1
tzdata==2023.3
tzlocal==5.0.1
urllib3==2.0.4
waitress==2.1.2
Werkzeug==2.3.6
windows-curses==2.3.1
opentok>=3.0
